requirements.txt
openai-agents
openai
python-dotenv
rich
pypdf
python-docx

